+++ CHANGELOG.md
@@ -4,7 +4,8 @@ The following is a history of the changes made to this project.
 
 ## v2.8.5 *(in development)*
 
-* The ability to assign RTL-SDR devices to all decoder installation scripts has been added.
+* The option to install ACARSDEC and ACARSERV in now available as a decoder option.
+* The ability to assign devices to additional decoder installations have been added to all scripts.
 
 ## v2.8.5 *(July 23rd, 2024)*